Move NavBar and HomePage
