+++ app/routes.ts
@@ -1,5 +1,5 @@
 import { type RouteConfig, index } from "@react-router/dev/routes";
 
 export default [
-	index("pages/home.tsx")
+	index("./home.tsx"),
 ] satisfies RouteConfig;
